requirements.txt
Django==4.1.13
djangorestframework==3.14.0
pymongo==3.12.3
djongo==1.3.6
pytz==2019.3
python-dotenv